+++ R/utils.R
@@ -1,10 +1,6 @@
 
-#' @importFrom magrittr is_less_than
-
 which_max <- function(x, tolerance = .Machine$double.eps^0.5) {
-  abs(max(x) - x) %>%
-    is_less_than(tolerance) %>%
-    which()
+  which(abs(max(x) - x) < tolerance)
 }
 
 #' @importFrom utils installed.packages