+++ inst/NEWS.md
@@ -1,4 +1,17 @@
 
+# 2.0.0
+
+* Add a `sparkline` class.
+
+* Remove the `print` option, the `sparkline` class auto-prints.
+
+* Remove the `magrittr` dependency.
+
+* Import symbols from the `clisymbols` package, to get automatic
+  fallback symbols in non-UTF8 terminals.
+
+* Better detection of UTF-8 terminals, use `l10n_info()`.
+
 # 1.0.1
 
 First version with a NEWS file.